requirements.txt
flask
opencv-python
ultralytics
pyserial